Refactor SPI related classes into new spi package
In anticipation of creating an SPI module
