+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/server/rest/RequestBuilder.java
@@ -16,10 +16,13 @@ package io.trino.s3.proxy.server.rest;
 import com.google.common.base.Splitter;
 import com.google.common.base.Suppliers;
 import io.airlift.log.Logger;
-import io.trino.s3.proxy.server.collections.ImmutableMultiMap;
-import io.trino.s3.proxy.server.collections.MultiMap;
-import io.trino.s3.proxy.server.rest.RequestContent.ContentType;
-import io.trino.s3.proxy.server.signing.RequestAuthorization;
+import io.trino.s3.proxy.spi.collections.ImmutableMultiMap;
+import io.trino.s3.proxy.spi.collections.MultiMap;
+import io.trino.s3.proxy.spi.rest.ParsedS3Request;
+import io.trino.s3.proxy.spi.rest.Request;
+import io.trino.s3.proxy.spi.rest.RequestContent;
+import io.trino.s3.proxy.spi.rest.RequestContent.ContentType;
+import io.trino.s3.proxy.spi.signing.RequestAuthorization;
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.UriBuilder;
