+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/server/rest/TrinoStsResource.java
@@ -17,14 +17,15 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import com.google.inject.Inject;
 import io.airlift.log.Logger;
-import io.trino.s3.proxy.server.collections.MultiMap;
-import io.trino.s3.proxy.server.credentials.AssumedRoleProvider;
-import io.trino.s3.proxy.server.credentials.EmulatedAssumedRole;
 import io.trino.s3.proxy.server.rest.AssumeRoleResponse.AssumeRoleResult;
 import io.trino.s3.proxy.server.rest.AssumeRoleResponse.AssumedRoleUser;
-import io.trino.s3.proxy.server.signing.SigningController;
-import io.trino.s3.proxy.server.signing.SigningMetadata;
-import io.trino.s3.proxy.server.signing.SigningServiceType;
+import io.trino.s3.proxy.spi.collections.MultiMap;
+import io.trino.s3.proxy.spi.credentials.AssumedRoleProvider;
+import io.trino.s3.proxy.spi.credentials.EmulatedAssumedRole;
+import io.trino.s3.proxy.spi.rest.Request;
+import io.trino.s3.proxy.spi.signing.SigningController;
+import io.trino.s3.proxy.spi.signing.SigningMetadata;
+import io.trino.s3.proxy.spi.signing.SigningServiceType;
 import jakarta.ws.rs.POST;
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Context;
