+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/spi/credentials/AssumedRoleProvider.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.credentials;
+package io.trino.s3.proxy.spi.credentials;
 
 import java.util.Optional;
