+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/spi/credentials/Credentials.java
@@ -11,9 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.credentials;
+package io.trino.s3.proxy.spi.credentials;
 
-import io.trino.s3.proxy.server.remote.RemoteSessionRole;
+import io.trino.s3.proxy.spi.remote.RemoteSessionRole;
 
 import java.util.Optional;
