+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/spi/rest/RequestContent.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.rest;
+package io.trino.s3.proxy.spi.rest;
 
 import java.io.InputStream;
 import java.util.Optional;
