+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/spi/signing/ChunkSigningSession.java
@@ -11,16 +11,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.credentials;
+package io.trino.s3.proxy.spi.signing;
 
-import java.util.Optional;
-
-public interface CredentialsProvider
+public interface ChunkSigningSession
 {
-    /**
-     * Return the credentials, if any, for the given access key and session.
-     * Your implementation should have a centralized credentials mechanism, likely
-     * some type of database along with a way of registering credentials, etc.
-     */
-    Optional<Credentials> credentials(String emulatedAccessKey, Optional<String> session);
+    void startChunk(String expectedSignature);
+
+    void complete();
+
+    void write(byte b);
+
+    void write(byte[] b, int off, int len);
 }
