+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/spi/signing/SigningContext.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.signing;
+package io.trino.s3.proxy.spi.signing;
 
 import java.util.Optional;
