+++ trino-s3-proxy/src/main/java/io/trino/s3/proxy/spi/signing/SigningMetadata.java
@@ -11,9 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.s3.proxy.server.signing;
+package io.trino.s3.proxy.spi.signing;
 
-import io.trino.s3.proxy.server.credentials.Credentials;
+import io.trino.s3.proxy.spi.credentials.Credentials;
 
 import java.util.Optional;
