+++ trino-s3-proxy/src/test/java/io/trino/s3/proxy/server/testing/TestingSecurityFacade.java
@@ -13,10 +13,10 @@
  */
 package io.trino.s3.proxy.server.testing;
 
-import io.trino.s3.proxy.server.rest.ParsedS3Request;
-import io.trino.s3.proxy.server.security.SecurityFacade;
-import io.trino.s3.proxy.server.security.SecurityFacadeProvider;
-import io.trino.s3.proxy.server.security.SecurityResponse;
+import io.trino.s3.proxy.spi.rest.ParsedS3Request;
+import io.trino.s3.proxy.spi.security.SecurityFacade;
+import io.trino.s3.proxy.spi.security.SecurityFacadeProvider;
+import io.trino.s3.proxy.spi.security.SecurityResponse;
 import jakarta.ws.rs.WebApplicationException;
 
 import java.util.concurrent.atomic.AtomicReference;
