+++ trino-s3-proxy/src/test/java/io/trino/s3/proxy/server/testing/TestingTrinoS3ProxyServerModule.java
@@ -20,7 +20,7 @@ import com.google.inject.Scopes;
 import io.trino.s3.proxy.server.TrinoS3ProxyModuleBuilder;
 import io.trino.s3.proxy.server.TrinoS3ProxyServerModule;
 import io.trino.s3.proxy.server.remote.RemoteS3Facade;
-import io.trino.s3.proxy.server.security.SecurityFacadeProvider;
+import io.trino.s3.proxy.spi.security.SecurityFacadeProvider;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
