+++ trino-s3-proxy/src/test/java/io/trino/s3/proxy/server/testing/containers/S3Container.java
@@ -18,9 +18,9 @@ import com.google.inject.BindingAnnotation;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import io.airlift.log.Logger;
-import io.trino.s3.proxy.server.credentials.Credential;
-import io.trino.s3.proxy.server.credentials.Credentials;
 import io.trino.s3.proxy.server.testing.TestingUtil.ForTesting;
+import io.trino.s3.proxy.spi.credentials.Credential;
+import io.trino.s3.proxy.spi.credentials.Credentials;
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import org.testcontainers.containers.Container;
