+++ trino-s3-proxy/src/test/resources/META-INF/services/io.trino.s3.proxy.spi.TrinoS3ProxyServerPlugin
